+++ js/ad-modal.js
@@ -1,0 +1,3 @@
+import './similar-ad';
+
+
