--- js/field.js
@@ -1,1 +1,0 @@
-import './map.js';
