+++ js/map.js
@@ -1,1 +1,1 @@
-import './popup.js';
+import './similar-ad.js';
